Goto table too large
The prior error, "Tokenizer tables too big to represent with 16-bit offsets," is from the generator's toArray function in token.ts. In an attempt to resolve it, I tried removing universalAlpha outright, as well as character. This now yields the error "Goto table too large" from computeGotoTable in build.ts.